flurry
Struct HashSetRef

pub struct HashSetRef<'set, T, S = DefaultHashBuilder> { /* private fields */ }

A reference to a HashSet, constructed with HashSet::pin or HashSet::with_guard.

The current thread will be pinned for the duration of this reference. Keep in mind that this prevents the collection of garbage generated by the set.

Implementations

impl<T, S> HashSetRef<'_, T, S>

pub fn len(&self) -> usize

Returns the number of elements in the set.

See also HashSet::len.

pub fn is_empty(&self) -> bool

Returns true if the set is empty. Otherwise returns false.

See also HashSet::is_empty.

pub fn iter(&self) -> Keys<'_, T, ()>

An iterator visiting all elements in arbitrary order.

The iterator element type is &'g T.

See also HashSet::iter.

impl<T, S> HashSetRef<'_, T, S>
where T: Hash + Ord, S: BuildHasher,

pub fn contains<Q>(&self, value: &Q) -> bool
where T: Borrow<Q>, Q: ?Sized + Hash + Ord,

Returns true if the given value is an element of this set.

See also HashSet::contains.

pub fn get<'g, Q>(&'g self, value: &Q) -> Option<&'g T>
where T: Borrow<Q>, Q: ?Sized + Hash + Ord,

Returns a reference to the element in the set, if any, that is equal to the given value.

See also HashSet::get.

pub fn is_disjoint(&self, other: &HashSetRef<'_, T, S>) -> bool

Returns true if self has no elements in common with other.

See also HashSet::is_disjoint.

pub fn is_subset(&self, other: &HashSetRef<'_, T, S>) -> bool

Returns true if the set is a subset of another, i.e., other contains at least all the values in self.

See also HashSet::is_subset.

pub fn is_superset(&self, other: &HashSetRef<'_, T, S>) -> bool

Returns true if the set is a superset of another, i.e., self contains at least all the values in other.

See also HashSet::is_superset.

impl<T, S> HashSetRef<'_, T, S>
where T: Sync + Send + Clone + Hash + Ord, S: BuildHasher,

pub fn insert(&self, value: T) -> bool

Adds a value to the set.

See also HashSet::insert.

pub fn remove<Q>(&self, value: &Q) -> bool
where T: Borrow<Q>, Q: ?Sized + Hash + Ord,

Removes a value from the set.

See also HashSet::remove.

pub fn take<'g, Q>(&'g self, value: &Q) -> Option<&'g T>
where T: Borrow<Q>, Q: ?Sized + Hash + Ord,

Removes and returns the value in the set, if any, that is equal to the given one.

See also HashSet::take.

pub fn retain<F>(&self, f: F)
where F: FnMut(&T) -> bool,

Retains only the elements specified by the predicate.

See also HashSet::retain.

impl<T, S> HashSetRef<'_, T, S>
where T: Clone + Ord,

pub fn clear(&self)

Clears the set, removing all elements.

See also HashSet::clear.

pub fn reserve(&self, additional: usize)

Tries to reserve capacity for at least additional more elements to be inserted into the underlying HashSet.

See also HashSet::reserve.